refined

+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: monoue <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/26 16:01:19 by monoue            #+#    #+#             */
-/*   Updated: 2020/08/20 10:31:09 by monoue           ###   ########.fr       */
+/*   Updated: 2020/08/20 11:59:46 by monoue           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		s1_index;
-	int		s2_index;
+	size_t	s1_index;
+	size_t	s2_index;
 	char	*joined_str;
 
 	if (s1 == NULL)
